refactor: split runner.rs god-file into modules
