+++ homeboy.json
@@ -509,9 +509,8 @@
     "audit": {
       "context_id": "homeboy",
       "created_at": "2026-06-22T13:13:12Z",
-      "item_count": 341,
+      "item_count": 338,
       "known_fingerprints": [
-        "compiler::src/commands/runner.rs::CompilerWarning",
         "compiler::src/commands/runs/gh_actions.rs::CompilerWarning",
         "compiler::src/commands/trace.rs::CompilerWarning",
         "compiler::src/core/agent_runtime_manifest.rs::CompilerWarning",
@@ -744,8 +743,6 @@
         "structural::src/commands/observe.rs::HighItemCount",
         "structural::src/commands/report/browser_evidence_compare.rs::GodFile",
         "structural::src/commands/report/failure_digest.rs::HighItemCount",
-        "structural::src/commands/runner.rs::GodFile",
-        "structural::src/commands/runner.rs::HighItemCount",
         "structural::src/commands/runner/doctor.rs::GodFile",
         "structural::src/commands/runs/gh_actions.rs::HighItemCount",
         "structural::src/commands::DirectorySprawl",
