+++ src/commands/runner/registry.rs
@@ -1,0 +1,272 @@
+use std::collections::HashMap;
+
+use serde_json::Value;
+
+use homeboy::core::redaction::RedactionPolicy;
+use homeboy::core::runners::{self as runner, ReverseRunnerConnectOptions, Runner, RunnerKind};
+use homeboy::core::server::{RunnerPolicy, RunnerSettings};
+use homeboy::core::MergeOutput;
+
+use super::super::{CmdResult, DynamicSetArgs};
+use super::cli::RunnerKindArg;
+use super::types::{RunnerConnectionOutput, RunnerExtra, RunnerOutput, REDACTED_ENV_VALUE};
+
+pub(super) struct RunnerAddInput {
+    pub(super) json: Option<String>,
+    pub(super) skip_existing: bool,
+    pub(super) id: Option<String>,
+    pub(super) kind: Option<RunnerKindArg>,
+    pub(super) server: Option<String>,
+    pub(super) workspace_root: Option<String>,
+    pub(super) settings: RunnerSettings,
+}
+
+pub(super) fn add(input: RunnerAddInput) -> CmdResult<RunnerOutput> {
+    let json_spec = if let Some(spec) = input.json {
+        spec
+    } else {
+        let id = input.id.ok_or_else(|| {
+            homeboy::core::Error::validation_invalid_argument(
+                "id",
+                "Missing required argument: id",
+                None,
+                None,
+            )
+        })?;
+        let kind = input.kind.map(RunnerKind::from).unwrap_or_else(|| {
+            if input.server.is_some() {
+                RunnerKind::Ssh
+            } else {
+                RunnerKind::Local
+            }
+        });
+        let new_runner = Runner {
+            id,
+            kind,
+            server_id: input.server,
+            workspace_root: input.workspace_root,
+            settings: input.settings,
+            env: HashMap::new(),
+            secret_env: HashMap::new(),
+            resources: HashMap::<String, Value>::new(),
+            policy: RunnerPolicy::default(),
+        };
+
+        homeboy::core::config::to_json_string(&new_runner)?
+    };
+
+    match runner::create(&json_spec, input.skip_existing)? {
+        homeboy::core::CreateOutput::Single(result) => Ok((
+            RunnerOutput {
+                command: "runner.add".to_string(),
+                id: Some(result.id),
+                entity: Some(result.entity),
+                updated_fields: vec!["created".to_string()],
+                ..Default::default()
+            },
+            0,
+        )),
+        homeboy::core::CreateOutput::Bulk(summary) => {
+            let exit_code = summary.exit_code();
+            Ok((
+                RunnerOutput {
+                    command: "runner.add".to_string(),
+                    import: Some(summary),
+                    ..Default::default()
+                },
+                exit_code,
+            ))
+        }
+    }
+}
+
+pub(super) fn list() -> CmdResult<RunnerOutput> {
+    Ok((
+        RunnerOutput {
+            command: "runner.list".to_string(),
+            entities: runner::list()?,
+            extra: RunnerExtra {
+                sessions: runner::statuses()?,
+                ..Default::default()
+            },
+            ..Default::default()
+        },
+        0,
+    ))
+}
+
+pub(super) fn enable(
+    server_id: &str,
+    workspace_root: Option<String>,
+    settings: RunnerSettings,
+) -> CmdResult<RunnerOutput> {
+    let mut spec = serde_json::Map::new();
+    if let Some(workspace_root) = workspace_root {
+        spec.insert("workspace_root".to_string(), workspace_root.into());
+    }
+    if let Some(homeboy_path) = settings.homeboy_path {
+        spec.insert("homeboy_path".to_string(), homeboy_path.into());
+    }
+    if settings.daemon {
+        spec.insert("daemon".to_string(), true.into());
+    }
+    if let Some(concurrency_limit) = settings.concurrency_limit {
+        spec.insert("concurrency_limit".to_string(), concurrency_limit.into());
+    }
+    if let Some(artifact_policy) = settings.artifact_policy {
+        spec.insert("artifact_policy".to_string(), artifact_policy.into());
+    }
+    let runner = runner::enable_server_runner(server_id, Value::Object(spec))?;
+    Ok((
+        RunnerOutput {
+            command: "runner.enable".to_string(),
+            id: Some(runner.id.clone()),
+            entity: Some(runner),
+            updated_fields: vec!["runner".to_string()],
+            ..Default::default()
+        },
+        0,
+    ))
+}
+
+pub(super) fn show(id: &str) -> CmdResult<RunnerOutput> {
+    let runner = runner::load(id)?;
+    Ok((
+        RunnerOutput {
+            command: "runner.show".to_string(),
+            id: Some(runner.id.clone()),
+            entity: Some(runner),
+            ..Default::default()
+        },
+        0,
+    ))
+}
+
+pub(super) fn set(args: DynamicSetArgs) -> CmdResult<RunnerOutput> {
+    let merged = super::super::merge_dynamic_args(&args)?.ok_or_else(|| {
+        homeboy::core::Error::validation_invalid_argument(
+            "spec",
+            "Provide --json '<object>' or --base64 <encoded-json>",
+            None,
+            Some(vec![
+                "Arbitrary runner updates must use explicit JSON input.".to_string(),
+                "Example: homeboy runner set <id> --json '{\"workspace_root\":\"/srv/homeboy\"}'"
+                    .to_string(),
+            ]),
+        )
+    })?;
+    let (json_string, replace_fields) = super::super::finalize_set_spec(&merged, &args.replace)?;
+
+    match runner::merge(args.id.as_deref(), &json_string, &replace_fields)? {
+        MergeOutput::Single(result) => {
+            let entity = runner::load(&result.id)?;
+            Ok((
+                RunnerOutput {
+                    command: "runner.set".to_string(),
+                    id: Some(result.id),
+                    entity: Some(entity),
+                    updated_fields: result.updated_fields,
+                    ..Default::default()
+                },
+                0,
+            ))
+        }
+        MergeOutput::Bulk(summary) => {
+            let exit_code = summary.exit_code();
+            Ok((
+                RunnerOutput {
+                    command: "runner.set".to_string(),
+                    batch: Some(summary),
+                    ..Default::default()
+                },
+                exit_code,
+            ))
+        }
+    }
+}
+
+pub(super) fn remove(id: &str) -> CmdResult<RunnerOutput> {
+    runner::delete_safe(id)?;
+    Ok((
+        RunnerOutput {
+            command: "runner.remove".to_string(),
+            id: Some(id.to_string()),
+            deleted: vec![id.to_string()],
+            ..Default::default()
+        },
+        0,
+    ))
+}
+
+pub(super) fn connect(
+    id: &str,
+    reverse: bool,
+    runner_id: Option<String>,
+    broker_url: Option<String>,
+) -> CmdResult<RunnerOutput> {
+    let (report, exit_code) = if reverse {
+        let runner_id = runner_id.ok_or_else(|| {
+            homeboy::core::Error::validation_invalid_argument(
+                "runner",
+                "Provide --reverse-runner <runner-id> when using --reverse",
+                None,
+                None,
+            )
+        })?;
+        runner::connect_reverse(ReverseRunnerConnectOptions {
+            controller_id: id.to_string(),
+            runner_id,
+            broker_url,
+        })?
+    } else {
+        runner::connect(id)?
+    };
+    Ok((
+        RunnerOutput {
+            command: "runner.connect".to_string(),
+            id: Some(report.runner_id.clone()),
+            extra: RunnerExtra {
+                connection: Some(RunnerConnectionOutput::Connect(report)),
+                ..Default::default()
+            },
+            ..Default::default()
+        },
+        exit_code,
+    ))
+}
+
+pub(super) fn disconnect(id: &str) -> CmdResult<RunnerOutput> {
+    Ok((
+        RunnerOutput {
+            command: "runner.disconnect".to_string(),
+            id: Some(id.to_string()),
+            extra: RunnerExtra {
+                connection: Some(RunnerConnectionOutput::Disconnect(runner::disconnect(id)?)),
+                ..Default::default()
+            },
+            ..Default::default()
+        },
+        0,
+    ))
+}
+
+pub(super) fn redact_runner_output_env(output: &mut RunnerOutput) {
+    if let Some(runner) = output.entity.as_mut() {
+        redact_runner_env(runner);
+    }
+
+    for runner in &mut output.entities {
+        redact_runner_env(runner);
+    }
+}
+
+fn redact_runner_env(runner: &mut Runner) {
+    let policy = RedactionPolicy::default();
+    for (key, value) in runner.env.iter_mut() {
+        if policy.is_sensitive_key(key) {
+            *value = REDACTED_ENV_VALUE.to_string();
+        } else {
+            *value = policy.redact_string(value);
+        }
+    }
+}
